+++ libraries/lps/source/disjointness_checker.cpp
@@ -20,8 +20,8 @@ using namespace mcrl2::core::detail;
 
 // Auxiliary functions ----------------------------------------------------------------------------
 
-  bool disjoint_sets(bool* a_set_1, bool* a_set_2, int a_number_of_elements) {
-    for (int i = 0; i < a_number_of_elements; i++) {
+  bool disjoint_sets(bool* a_set_1, bool* a_set_2, size_t a_number_of_elements) {
+    for (size_t i = 0; i < a_number_of_elements; i++) {
       if (a_set_1[i] && a_set_2[i]) {
         return false;
       }
@@ -32,7 +32,7 @@ using namespace mcrl2::core::detail;
 // Class Disjointness_Checker ---------------------------------------------------------------------
   // Class Disjointness_Checker - Functions declared private --------------------------------------
 
-    void Disjointness_Checker::process_data_expression(int a_summand_number, ATermAppl a_expression) {
+    void Disjointness_Checker::process_data_expression(size_t a_summand_number, ATermAppl a_expression) {
       if (gsIsDataVarId(a_expression)) {
         size_t v_variable_index = ATindexedSetGetIndex(f_parameter_set, (ATerm) a_expression);
         if (v_variable_index != NON_EXISTING) {
@@ -60,7 +60,7 @@ using namespace mcrl2::core::detail;
 
     // --------------------------------------------------------------------------------------------
 
-    void Disjointness_Checker::process_multi_action(int a_summand_number, ATermAppl v_multi_action) {
+    void Disjointness_Checker::process_multi_action(size_t a_summand_number, ATermAppl v_multi_action) {
       ATermList v_actions = ATLgetArgument(v_multi_action, 0);
       ATermAppl v_action;
       ATermList v_expressions;
@@ -80,7 +80,7 @@ using namespace mcrl2::core::detail;
 
     // --------------------------------------------------------------------------------------------
 
-    void Disjointness_Checker::process_summand(int a_summand_number, ATermAppl a_summand) {
+    void Disjointness_Checker::process_summand(size_t a_summand_number, ATermAppl a_summand) {
       ATermAppl v_condition = ATAgetArgument(a_summand, 1);
       ATermAppl v_multi_action_or_delta = ATAgetArgument(a_summand, 2);
       ATermAppl v_time = ATAgetArgument(a_summand, 3);
@@ -140,7 +140,7 @@ using namespace mcrl2::core::detail;
       ATermList v_parameters = ATLgetArgument(a_process_equation, 0);
       ATermList v_summands = ATLgetArgument(a_process_equation, 1);
       ATermAppl v_summand;
-      int v_summand_number = 1;
+      size_t v_summand_number = 1;
 
       f_parameter_set = ATindexedSetCreate(ATgetLength(v_parameters) * 2, 50);
       f_number_of_summands = ATgetLength(v_summands);
@@ -167,7 +167,7 @@ using namespace mcrl2::core::detail;
 
     // --------------------------------------------------------------------------------------------
 
-    bool Disjointness_Checker::disjoint(int a_summand_number_1, int a_summand_number_2) {
+    bool Disjointness_Checker::disjoint(size_t a_summand_number_1, size_t a_summand_number_2) {
       if ((a_summand_number_1 <= f_number_of_summands) && (a_summand_number_2 <= f_number_of_summands)) {
         bool v_used_1_changed_2 = disjoint_sets(
                                &f_used_parameters_per_summand[a_summand_number_1 * f_number_of_parameters],
